+++ webpack/app/senaite.core.js
@@ -75,4 +75,52 @@ document.addEventListener("DOMContentLoaded", () => {
     }
   });
 
+
+  // BBB: create form Bootstrap navigation tabs for all fieldsets that are
+  //      located in a form with the CSS class "enableFormTabbing"
+  document.querySelectorAll("form.enableFormTabbing").forEach(function (form) {
+      let fieldsets = form.querySelectorAll("fieldset");
+      if (fieldsets.length === 0) return;
+
+      let nav = document.createElement("ul");
+      nav.className = "nav nav-tabs";
+      nav.setAttribute("role", "tablist");
+
+      let tabContent = document.createElement("div");
+      tabContent.className = "tab-content";
+
+      fieldsets.forEach((fieldset, index) => {
+        let legend = fieldset.querySelector("legend");
+        let tabId = "tab-" + index;
+
+        let li = document.createElement("li");
+        li.className = "nav-item";
+
+        let a = document.createElement("a");
+        a.className = "nav-link" + (index === 0 ? " active" : "");
+        a.setAttribute("data-toggle", "tab");
+        a.href = "#" + tabId;
+        a.setAttribute("role", "tab");
+        a.innerText = legend ? legend.innerText : "Tab " + (index + 1);
+
+        // remove the legend
+        legend.remove();
+
+        li.appendChild(a);
+        nav.appendChild(li);
+
+        let tabPane = document.createElement("div");
+        tabPane.className = "tab-pane fade" + (index === 0 ? " show active" : "");
+        tabPane.id = tabId;
+        tabPane.setAttribute("role", "tabpanel");
+        fieldset.parentNode.insertBefore(tabPane, fieldset);
+        tabPane.appendChild(fieldset);
+
+        tabContent.appendChild(tabPane);
+      });
+
+      form.insertBefore(nav, form.firstChild);
+      form.insertBefore(tabContent, form.firstChild.nextSibling);
+    });
+
 });